pyproject.toml
[tool.poetry]
name = "reminder"
version = "0.1.0"
description = ""
authors = ["changiinlee <changjin9792@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
sqlalchemy = "^2.0.25"
python-dotenv = "^1.0.1"
aiomysql = "^0.2.0"
python-multipart = "^0.0.7"
openai = "^1.11.1"
boto3 = "^1.34.36"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
